Final adjustments
Added missing docstrings, adjusted formatting, and removed redundant file.
