+++ Group_6_Assignment_1/Question_1.py
