+++ Group_6_Assignment_1/Question_2.py
